Enable the in-game cash shop on Android and fix its wire-protocol bugs
Turns on the GameShop subsystem for Android (previously #ifdef'd out
entirely) and wires the mobile shop button to actually open it, mirroring
the PC "X" toolbar handler's script-load sequence. Content loads from
locally-shipped IBSCategory/Package/Product.txt - the shop's downloader
already skips its WinINet-only live fetch whenever local files exist, so
the Android build never needs a working FTP/HTTP client; the downloader
classes are stubbed to fail gracefully instead of linking WinINet.

Three separate, real bugs surfaced once this code actually compiled and
ran on Android for the first time (all pre-existing in the shared source,
just never exercised there before):

- CMsgBoxIGSBuyPackageItem::Initialize() never set m_iPriceSeq, so every
  single-price ("Buy" without a picker) purchase sent PriceSeq 0 instead
  of the item's real price sequence.
- Several PMSG_CASHSHOP_* wire structs (WSclient.h) and the matching
  SendRequestIGS_* macros (wsclientinline.h) declared their sequence-ID
  fields as bare `long`. That's 4 bytes on Windows but 8 on Android's
  64-bit ABI, so every packed struct/packet using it silently
  desynchronized in flight on Android only - explaining both "sold out"
  on an otherwise-valid buy request and the storage list staying empty
  after a purchase actually succeeded. Switched to `int`, which is a
  fixed 4 bytes on both platforms and matches the wire format PC already
  used.

Also fixes several Android-only compile blockers hit while bringing up
this previously-never-built-for-Android code (bare wide-char Win32 API
calls needing ANSI shims added to PlatformDefs.h, ifstream paths needing
backslash normalization, a void-returning ShellExecute shim, stray MSVC
extra-qualification syntax) - all gated to Android or otherwise inert on
PC, verified via a clean PC rebuild.

+++ Source_Server-Mobi/5.Main/source/WSclient.h
@@ -3293,9 +3293,9 @@ typedef struct
 {
 	PBMSG_HEADER2		h;
 	
-	long				lBuyItemPackageSeq;
-	long				lBuyItemDisplaySeq;
-	long				lBuyItemPriceSeq;
+	int				lBuyItemPackageSeq;
+	int				lBuyItemDisplaySeq;
+	int				lBuyItemPriceSeq;
 	WORD				wItemCode;
 }PMSG_CASHSHOP_BUYITEM_REQ, *LPPMSG_CASHSHOP_BUYITEM_REQ;
 
@@ -3307,7 +3307,7 @@ typedef struct
 	PBMSG_HEADER2		h;
 
 	BYTE				byResultCode;
-	long				lItemLeftCount;
+	int				lItemLeftCount;
 }PMSG_CASHSHOP_BUYITEM_ANS, *LPPMSG_CASHSHOP_BUYITEM_ANS;
 
 //----------------------------------------------------------------------------
@@ -3317,10 +3317,10 @@ typedef struct
 {
 	PBMSG_HEADER2		h;
 
-	long				lGiftItemPackageSeq;
-	long				lDiftItemDisplaySeq;
-	long				lGiftItemPriceSeq;
-	long				lSaleZone;
+	int				lGiftItemPackageSeq;
+	int				lDiftItemDisplaySeq;
+	int				lGiftItemPriceSeq;
+	int				lSaleZone;
 	char				chReceiveUserID[MAX_ID_SIZE+1];
 	char				chMessage[MAX_GIFT_MESSAGE_SIZE];
 }PMSG_CASHSHOP_GIFTSEND_REQ, *LPPMSG_CASHSHOP_GIFTSEND_REQ;
@@ -3333,7 +3333,7 @@ typedef struct
 	PBMSG_HEADER2		h;
 
 	BYTE				byResultCode;
-	long				lItemLeftCount;
+	int				lItemLeftCount;
 	double				dLimitedCash;
 }PMSG_CASHSHOP_GIFTSEND_ANS, *LPPMSG_CASHSHOP_GIFTSEND_ANS;
 
@@ -3366,11 +3366,11 @@ typedef struct
 {
 	PBMSG_HEADER2		h;
 
-	long				lStorageIndex;
-	long				lItemSeq;
-	long				lStorageGroupCode;
-	long				lProductSeq;
-	long				lPriceSeq;
+	int				lStorageIndex;
+	int				lItemSeq;
+	int				lStorageGroupCode;
+	int				lProductSeq;
+	int				lPriceSeq;
 	double				dCashPoint;
 	char				chItemType;
 }PMSG_CASHSHOP_STORAGELIST, *LPPMSG_CASHSHOP_STORAGELIST;
@@ -3382,11 +3382,11 @@ typedef struct
 {
 	PBMSG_HEADER2		h;
 	
-	long				lStorageIndex;
-	long				lItemSeq;
-	long				lStorageGroupCode;
-	long				lProductSeq;
-	long				lPriceSeq;
+	int				lStorageIndex;
+	int				lItemSeq;
+	int				lStorageGroupCode;
+	int				lProductSeq;
+	int				lPriceSeq;
 	double				dCashPoint;
 	char				chItemType;
 
@@ -3448,7 +3448,7 @@ typedef struct
 {
 	PBMSG_HEADER2		h;
 
-	long				lPackageSeq;
+	int				lPackageSeq;
 }PMSG_CASHSHOP_ITEMBUY_LEFT_COUNT_REQ, *LPPMSG_CASHSHOP_ITEMBUY_LEFT_COUNT_REQ;
 
 //----------------------------------------------------------------------------
@@ -3458,8 +3458,8 @@ typedef struct
 {
 	PBMSG_HEADER2		h;
 
-	long				lPackageSeq;
-	long				lLeftCount;
+	int				lPackageSeq;
+	int				lLeftCount;
 }PMSG_CASHSHOP_ITEMBUY_LEFT_COUNT_ANS, *LPPMSG_CASHSHOP_ITEMBUY_LEFT_COUNT_ANS;
 
 //----------------------------------------------------------------------------
@@ -3469,8 +3469,8 @@ typedef struct
 {
 	PBMSG_HEADER2		h;
 
-	long				lStorageSeq;
-	long				lStorageItemSeq;
+	int				lStorageSeq;
+	int				lStorageItemSeq;
 	char				chStorageItemType;
 }PMSG_CASHSHOP_STORAGE_ITEM_THROW_REQ, *LPPMSG_CASHSHOP_STORAGE_ITEM_THROW_REQ;
 
@@ -3491,8 +3491,8 @@ typedef struct
 {
 	PBMSG_HEADER2		h;
 
-	long				lStorageSeq;
-	long				lStorageItemSeq;
+	int				lStorageSeq;
+	int				lStorageItemSeq;
 
 	WORD				wItemCode;
 	char				chStorageItemType;
@@ -3527,7 +3527,7 @@ typedef struct
 {
 	PBMSG_HEADER2			h;
 	
-	long				lEventCategorySeq;
+	int				lEventCategorySeq;
 }PMSG_CASHSHOP_EVENTITEM_REQ, *LPPMSG_CASHSHOP_EVENTITEM_REQ;
 
 
@@ -3546,7 +3546,7 @@ typedef struct
 {
 	PBMSG_HEADER2			h;
 	
-	long				lPackageSeq[INGAMESHOP_DISPLAY_ITEMLIST_SIZE];
+	int				lPackageSeq[INGAMESHOP_DISPLAY_ITEMLIST_SIZE];
 }PMSG_CASHSHOP_EVENTITEM_LIST, *LPPMSG_CASHSHOP_EVENTITEM_LIST;
 
 typedef struct
